BACKEND/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
google-generativeai
faiss-cpu
numpy
pymupdf
python-dotenv
playwright